pyproject.toml
[tool.poetry]
name = "fast-api-test"
version = "0.1.0"
description = ""
authors = ["Jacob Shore <jshore@datalign.net>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.63.0"}
Hypercorn = "^0.11.2"
pony = "^0.7.14"
psycopg2 = "^2.8.6"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
